size-predictor/server/requirements.txt
flask
flask_cors
flask_jwt_extended
flask_bcrypt
python-dotenv
datetime
pymongo
black